+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.1.0
+
+- fix: resizing canvas by element size and not the window size
+- feat: passing canvas element or canvas id to constructor
+
 # 1.0.1
 
 - fix: resize